requirements.txt
Flask==3.0.0
flask-cors==4.0.0
PyMuPDF>=1.24.2
Werkzeug==3.0.1
pytesseract==0.3.10
Pillow==10.1.0
